IQM loading works

+++ src/com/raylib/java/models/iqm/IQMMesh.java
@@ -5,7 +5,7 @@ public class IQMMesh {
     public int name;
     public int material;
     public int first_vertex, num_vertexes;
-    public long first_triangle, num_triangles;
+    public int first_triangle, num_triangles;
 
     public IQMMesh() {
